Admin

+++ Webdock-admin/src/index.ts
@@ -27,6 +27,8 @@ import {
   UpdateStatus,
   DeleteStatus,
 } from "./querys/statusQuerys";
+import { IsAdministrator } from "../../shared/services/Autherization";
+import { IAdmin } from "../../shared/services/interfaces/IAdministrator";
 
 const app = new Elysia()
   .get("/status", () => {
@@ -36,30 +38,89 @@ const app = new Elysia()
   })
   .group("/Users", (app) =>
     app
-      .get("/", async () => {
+      .get("/", async ({ request: { headers } }) => {
+        const requesterId = headers.get("requesterId");
+        const isAdministrator = await IsAdministrator({
+          requesterId,
+        } as IAdmin);
+        if (!isAdministrator) {
+          throw new Error("Unauthorized");
+        }
+
         return await GetAllUsers();
       })
-      .get("/:id", async ({ params: { id } }) => {
+      .get("/:id", async ({ params: { id }, request: { headers } }) => {
+        const requesterId = headers.get("requesterId");
+        const isAdministrator = await IsAdministrator({
+          requesterId,
+        } as IAdmin);
+        if (!isAdministrator) {
+          throw new Error("Unauthorized");
+        }
+
         return await GetUser({ id } as IGetUser);
       })
-      .put("/role/:id", async ({ params: { id }, body }) => {
-        const { role } = body as IUpdateUserRole;
-        return await UpdateUserRole({ id, role } as IUpdateUserRole);
-      })
+      .put(
+        "/role/:id",
+        async ({ params: { id }, body, request: { headers } }) => {
+          const requesterId = headers.get("requesterId");
+          const isAdministrator = await IsAdministrator({
+            requesterId,
+          } as IAdmin);
+          if (!isAdministrator) {
+            throw new Error("Unauthorized");
+          }
+
+          const { role } = body as IUpdateUserRole;
+          return await UpdateUserRole({ id, role } as IUpdateUserRole);
+        }
+      )
   )
   .group("/Categories", (app) =>
     app
-      .get("/", async () => {
+      .get("/", async ({ request: { headers } }) => {
+        const requesterId = headers.get("requesterId");
+        const isAdministrator = await IsAdministrator({
+          requesterId,
+        } as IAdmin);
+        if (!isAdministrator) {
+          throw new Error("Unauthorized");
+        }
+
         return await GetAllCategories();
       })
-      .get("/:id", async ({ params: { id } }) => {
+      .get("/:id", async ({ params: { id }, request: { headers } }) => {
+        const requesterId = headers.get("requesterId");
+        const isAdministrator = await IsAdministrator({
+          requesterId,
+        } as IAdmin);
+        if (!isAdministrator) {
+          throw new Error("Unauthorized");
+        }
+
         return await GetCategory({ id: Number(id) } as IGetCategory);
       })
-      .post("/", async ({ body }) => {
+      .post("/", async ({ body, request: { headers } }) => {
+        const requesterId = headers.get("requesterId");
+        const isAdministrator = await IsAdministrator({
+          requesterId,
+        } as IAdmin);
+        if (!isAdministrator) {
+          throw new Error("Unauthorized");
+        }
+
         const { name } = body as ICreateCategory;
         return await CreateCategory({ name } as ICreateCategory);
       })
-      .put("/:id", async ({ params: { id }, body }) => {
+      .put("/:id", async ({ params: { id }, body, request: { headers } }) => {
+        const requesterId = headers.get("requesterId");
+        const isAdministrator = await IsAdministrator({
+          requesterId,
+        } as IAdmin);
+        if (!isAdministrator) {
+          throw new Error("Unauthorized");
+        }
+
         const { name } = body as IUpdateCategory;
         const update = {
           id: Number(id),
@@ -68,23 +129,63 @@ const app = new Elysia()
 
         return await UpdateCategory(update as IUpdateCategory);
       })
-      .delete("/:id", async ({ params: { id } }) => {
+      .delete("/:id", async ({ params: { id }, request: { headers } }) => {
+        const requesterId = headers.get("requesterId");
+        const isAdministrator = await IsAdministrator({
+          requesterId,
+        } as IAdmin);
+        if (!isAdministrator) {
+          throw new Error("Unauthorized");
+        }
+
         return await DeleteCategory({ id: Number(id) } as IDeleteCategory);
       })
   )
   .group("/Statuses", (app) =>
     app
-      .get("/", async () => {
+      .get("/", async ({ request: { headers } }) => {
+        const requesterId = headers.get("requesterId");
+        const isAdministrator = await IsAdministrator({
+          requesterId,
+        } as IAdmin);
+        if (!isAdministrator) {
+          throw new Error("Unauthorized");
+        }
+
         return await GetAllStatuses();
       })
-      .get("/:id", async ({ params: { id } }) => {
+      .get("/:id", async ({ params: { id }, request: { headers } }) => {
+        const requesterId = headers.get("requesterId");
+        const isAdministrator = await IsAdministrator({
+          requesterId,
+        } as IAdmin);
+        if (!isAdministrator) {
+          throw new Error("Unauthorized");
+        }
+
         return await GetStatus({ id: Number(id) } as IGetStatus);
       })
-      .post("/", async ({ body }) => {
+      .post("/", async ({ body, request }) => {
+        const requesterId = request.headers.get("requesterId");
+        const isAdministrator = await IsAdministrator({
+          requesterId,
+        } as IAdmin);
+        if (!isAdministrator) {
+          throw new Error("Unauthorized");
+        }
+
         const { name } = body as ICreateStatus;
         return await CreateStatus({ name } as ICreateStatus);
       })
-      .put("/:id", async ({ params: { id }, body }) => {
+      .put("/:id", async ({ params: { id }, body, request: { headers } }) => {
+        const requesterId = headers.get("requesterId");
+        const isAdministrator = await IsAdministrator({
+          requesterId,
+        } as IAdmin);
+        if (!isAdministrator) {
+          throw new Error("Unauthorized");
+        }
+
         const { name } = body as IUpdateStatus;
         const update = {
           id: Number(id),
@@ -93,7 +194,15 @@ const app = new Elysia()
 
         return await UpdateStatus(update as IUpdateStatus);
       })
-      .delete("/:id", async ({ params: { id } }) => {
+      .delete("/:id", async ({ params: { id }, request: { headers } }) => {
+        const requesterId = headers.get("requesterId");
+        const isAdministrator = await IsAdministrator({
+          requesterId,
+        } as IAdmin);
+        if (!isAdministrator) {
+          throw new Error("Unauthorized");
+        }
+
         return await DeleteStatus({ id: Number(id) } as IDeleteStatus);
       })
   )
